+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - `Innmind\BlackBox\Runner\Assert::matches()`
 - `Innmind\BlackBox\Runner\Assert::time()`
 - `Innmind\BlackBox\Runner\Assert::memory()`
+- `Innmind\BlackBox\Set\Slice`
 
 ### Fixed