+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-# [Unreleased]
+# 5.3.0 - 2023-08-27
 
 ### Added